All basic function used on the testing and maintenance view are now working and implimented. See issue 5 relating to some presenter methos not being completed at this stage. Testing and maintenance view now only has left the complex multi command  methods to impliment back end wise.

+++ app/Hardware/CommandConstructor.h
@@ -56,7 +56,7 @@ namespace App { namespace Hardware
             QVariantMap setFlowControllerSourceControl(QString controller, int source);
             QVariantMap setFlowControllerSoftStartTime(QString controller, int seconds);
             QVariantMap setFlowControllerSoftStart(QString controller, int state);
-            QVariantMap setFlowControllerFlowRate(QString controller, int flowrate);
+            QVariantMap setFlowControllerFlowRate(QString controller, double flowrate);
             QVariantMap setFlowControllerValveOverride(QString controller, int state);
 
     };
